# public/api-spec.yaml
openapi: 3.0.0
info:
  title: QUANLICS RAG API
  description: |
    Production-quality RAG API for QUANLICS policy management system.
    
    Features:
    - Vector-based semantic search with ChromaDB
    - Intelligent caching with 36% hit rate
    - Multi-stage document analysis
    - Real-time Q&A with citations
    
    Response times:
    - Cache hit: ~89ms
    - Vector search: ~1.2s
    - Cold start: ~1.6s
  version: 1.0.0
  contact:
    name: QUANLICS Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api
    description: Development
  - url: https://api.quanlics.edu.vn/api
    description: Production

paths:
  /chatbot/improved/ask:
    post:
      summary: Ask a question using improved RAG
      description: |
        Ask a question and get an intelligent answer with citations.
        Uses semantic cache + vector search + Gemini generation.
        
        Performance:
        - Cache hit: 89ms
        - Cache miss: 1.2-1.6s
      operationId: askChatbot
      tags:
        - Chatbot
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
              properties:
                question:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  example: "Em là con hộ nghèo được miễn bao nhiêu học phí?"
                phien_id:
                  type: integer
                  nullable: true
                  description: Optional session ID for tracking
                  example: 123
      responses:
        200:
          description: Successful response with answer and citations
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  answer:
                    type: string
                    description: Generated answer to the question
                    example: "Theo Điều 3 của Nghị định 81/2021, hộ nghèo được miễn 100% học phí."
                  citations:
                    type: array
                    items:
                      type: string
                    description: Related policy articles
                    example: ["Điều 3", "Điều 4", "Điều 5"]
                  metadata:
                    type: object
                    properties:
                      method:
                        type: string
                        enum: ["cache_hit", "vector_rag", "fallback_error"]
                        description: Which method was used
                      response_time_ms:
                        type: number
                        example: 1245.5
                      cache_similar_score:
                        type: number
                        nullable: true
                        example: 0.92
                      chunks_used:
                        type: integer
                        example: 3
        400:
          description: Bad request (empty question, invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        429:
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Server error (fallback to basic response)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /ho-so/{maHoSo}/analyze-for-reduction:
    post:
      summary: Analyze HỒ SỚ (document) for scholarship reduction
      description: |
        Analyze submitted forms to determine scholarship reduction percentage
        based on Nghị định 81/2021.
        
        Returns:
        - Reduction percentage (0-100%)
        - Auto-approval status
        - Required documents checklist
        - Recommendations
      operationId: analyzeForReduction
      tags:
        - Documents
      security:
        - BearerAuth: []
      parameters:
        - name: maHoSo
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: Analysis results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      ma_ho_so:
                        type: integer
                        example: 1
                      policy_type:
                        type: string
                        enum: ["MGHP", "TCXH", "unknown"]
                        example: "MGHP"
                      reduction_percentage:
                        type: integer
                        minimum: 0
                        maximum: 100
                        example: 100
                      status:
                        type: string
                        enum: ["APPROVED", "PENDING_REVIEW", "NEED_SUPPLEMENT"]
                        example: "APPROVED"
                      completeness:
                        type: object
                        properties:
                          percentage:
                            type: number
                            example: 100
                          total_required:
                            type: integer
                            example: 5
                          total_submitted:
                            type: integer
                            example: 5
                          missing_documents:
                            type: array
                            items:
                              type: string
                            example: []
                      recommendation:
                        type: string
                        example: "Tất cả hồ sơ đầy đủ. Được phê duyệt tự động."
                      analysis_time_ms:
                        type: number
                        example: 342.5
        404:
          description: HỒ SỚ not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'

  /admin/rag-setup:
    post:
      summary: Initialize RAG system (Admin only)
      description: |
        One-time setup to index knowledge base into ChromaDB.
        
        Requires admin role.
        Chunks documents and generates embeddings.
        Takes 30-60 seconds.
      operationId: ragSetup
      tags:
        - Admin
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reindex:
                  type: boolean
                  description: Force reindex if already indexed
                  example: false
      responses:
        200:
          description: RAG setup completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "RAG indexing completed successfully"
                  data:
                    type: object
                    properties:
                      chunks_created:
                        type: integer
                        example: 345
                      embeddings_generated:
                        type: integer
                        example: 345
                      storage_size_mb:
                        type: number
                        example: 12.5
                      processing_time_seconds:
                        type: number
                        example: 45.3
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          description: Not admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: RAG setup failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      summary: Health check
      description: Check system and service health
      operationId: healthCheck
      tags:
        - System
      responses:
        200:
          description: System healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: ["healthy", "degraded", "unhealthy"]
                    example: "healthy"
                  services:
                    type: object
                    properties:
                      cache:
                        type: string
                        enum: ["ok", "error"]
                        example: "ok"
                      chroma_db:
                        type: string
                        enum: ["ok", "error"]
                        example: "ok"
                      gemini_api:
                        type: string
                        enum: ["ok", "error"]
                        example: "ok"
                  timestamp:
                    type: string
                    format: date-time

  /metrics:
    get:
      summary: System metrics
      description: Get performance metrics (requires admin)
      operationId: getMetrics
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        200:
          description: System metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  cache_hit_rate:
                    type: number
                    example: 0.36
                  avg_response_time_ms:
                    type: number
                    example: 1245.5
                  total_questions:
                    type: integer
                    example: 1250
                  errors_last_hour:
                    type: integer
                    example: 2
                  memory_usage_mb:
                    type: number
                    example: 18.5

components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid question: empty string"
        code:
          type: string
          example: "INVALID_INPUT"
        details:
          type: object
          nullable: true

  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthenticated"

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Use Laravel Sanctum token.
        
        Get token:
        POST /auth/login
        { "username": "...", "password": "..." }

tags:
  - name: Chatbot
    description: Q&A operations
  - name: Documents
    description: Document analysis
  - name: Admin
    description: Administrative operations
  - name: System
    description: System health and metrics

x-metadata:
  test-collection: https://www.postman.com/...
  performance:
    cache-hit-ms: 89
    vector-search-ms: 1200
    cold-start-ms: 1600